Move shared_linux to shared/linux
