+++ src/platform/backends/shared/linux/backend_utils.cpp
